♻️ refactor : 패키지 네이밍 변경 entity → model
