+++ src/main/java/com/devcourse/checkmoi/domain/token/repository/TokenRepository.java
@@ -1,10 +1,9 @@
 package com.devcourse.checkmoi.domain.token.repository;
 
 
-import com.devcourse.checkmoi.domain.token.entity.Token;
-import org.springframework.data.jpa.repository.JpaRepository;
-
+import com.devcourse.checkmoi.domain.token.model.Token;
 import java.util.Optional;
+import org.springframework.data.jpa.repository.JpaRepository;
 
 public interface TokenRepository extends JpaRepository<Token, Long> {
