+++ src/main/java/com/devcourse/checkmoi/domain/user/repository/UserRepository.java
@@ -1,11 +1,11 @@
 package com.devcourse.checkmoi.domain.user.repository;
 
+import com.devcourse.checkmoi.domain.user.model.User;
 import java.util.Optional;
-
-import com.devcourse.checkmoi.domain.user.entity.User;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 public interface UserRepository extends JpaRepository<User, Long> {
+
     Optional<User> findByOauthId(String oauthId);
 
 }
