+++ src/main/java/com/devcourse/checkmoi/global/security/oauth/UserProfile.java
@@ -1,13 +1,14 @@
 package com.devcourse.checkmoi.global.security.oauth;
 
-import com.devcourse.checkmoi.domain.user.entity.User;
-import com.devcourse.checkmoi.domain.user.entity.UserRole;
+import com.devcourse.checkmoi.domain.user.model.User;
+import com.devcourse.checkmoi.domain.user.model.UserRole;
 import com.devcourse.checkmoi.global.vo.Email;
 import lombok.Builder;
 import lombok.Getter;
 
 @Getter
 public class UserProfile {
+
     private final String oauthId;
     private final String provider;
     private final String name;
@@ -15,7 +16,8 @@ public class UserProfile {
     private final String profileImgUrl;
 
     @Builder
-    public UserProfile(String oauthId, String provider, String name, String email, String profileImgUrl) {
+    public UserProfile(String oauthId, String provider, String name, String email,
+        String profileImgUrl) {
         this.oauthId = oauthId;
         this.provider = provider;
         this.name = name;
